move semester into new module and add endpoint for getting semesters
