+++ packages/api-client/index.ts
@@ -26,6 +26,7 @@ import {
   ListInsightsResponse,
   DateRangeType,
   SubmitCourseParams,
+  SemesterPartial,
 } from "@koh/common";
 import Axios, { AxiosInstance, Method } from "axios";
 import { plainToClass } from "class-transformer";
@@ -188,6 +189,10 @@ class APIClient {
     delete: async () => this.req("GET", `/api/v1/seeds/delete`),
     create: async () => this.req("GET", `/api/v1/seeds/create`),
     fillQueue: async () => this.req("GET", `/api/v1/seeds/fill_queue`),
+  };
+  semesters = {
+    get: async (): Promise<SemesterPartial[]> =>
+      this.req("GET", `/api/v1/semesters`),
   };
   releaseNotes = {
     get: async (): Promise<GetReleaseNotesResponse> =>
