+++ packages/server/ormconfig.ts
@@ -2,7 +2,7 @@ import { config } from 'dotenv';
 import { AdminUserModel } from './src/admin/admin-user.entity';
 import { CourseModel } from './src/course/course.entity';
 import { OfficeHourModel } from './src/course/office-hour.entity';
-import { SemesterModel } from './src/course/semester.entity';
+import { SemesterModel } from './src/semester/semester.entity';
 import { CourseSectionMappingModel } from './src/login/course-section-mapping.entity';
 import { DesktopNotifModel } from './src/notification/desktop-notif.entity';
 import { PhoneNotifModel } from './src/notification/phone-notif.entity';
