+++ packages/server/src/course/course.entity.ts
@@ -13,7 +13,7 @@ import { EventModel } from '../profile/event-model.entity';
 import { UserCourseModel } from '../profile/user-course.entity';
 import { QueueModel } from '../queue/queue.entity';
 import { OfficeHourModel } from './office-hour.entity';
-import { SemesterModel } from './semester.entity';
+import { SemesterModel } from '../semester/semester.entity';
 import { AlertModel } from '../alerts/alerts.entity';
 
 /**
