+++ packages/server/src/semester/semester.controller.ts
@@ -1,0 +1,15 @@
+import { SemesterPartial } from '@koh/common';
+import { Controller, Get, UseGuards } from '@nestjs/common';
+import { JwtAuthGuard } from 'login/jwt-auth.guard';
+import { SemesterModel } from './semester.entity';
+
+@Controller('semesters')
+@UseGuards(JwtAuthGuard)
+export class SemesterController {
+  constructor() {}
+
+  @Get()
+  async get(): Promise<SemesterPartial[]> {
+    return SemesterModel.find();
+  }
+}
