+++ packages/server/src/semester/semester.module.ts
@@ -1,0 +1,7 @@
+import { Module } from '@nestjs/common';
+import { SemesterController } from './semester.controller';
+
+@Module({
+  controllers: [SemesterController],
+})
+export class SemesterModule {}
